backend/requirements.txt
# ================================================================
# POA Eventos — Python Dependencies (pinned)
# ================================================================

# --- Django core ---
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# --- GeoDjango / PostGIS ---
psycopg2-binary==2.9.9
djangorestframework-gis==1.0

# --- Authentication ---
dj-rest-auth==6.0.0
django-allauth==0.63.3
djangorestframework-simplejwt==5.3.1
cryptography==42.0.8

# --- Async tasks ---
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1

# --- Config / Env ---
django-environ==0.11.2

# --- Image processing ---
Pillow==10.3.0
python-magic==0.4.27

# --- HTTP / Utils ---
requests==2.32.2
python-slugify==8.0.4

# --- WSGI server ---
gunicorn==22.0.0

# --- Logging ---
structlog==24.1.0
django-structlog==8.0.0

# --- Rate limiting ---
django-ratelimit==4.1.0

# --- Storage (S3 optional) ---
django-storages==1.14.3
boto3==1.34.84

# --- Testing ---
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.2.0

# --- Linting ---
flake8==7.0.0
black==24.4.2
isort==5.13.2
